# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
#host: localhost
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /medicos:
    x-swagger-router-controller: epivector.medico.controller
    
    get:
      operationId: getMedicos
      description: Retorna la lista de entidades
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/MedicoList"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse"    
            
            
    post:
      description: Agrega una nueva entidad
      parameters:
        - in: body
          name: entidad
          schema:
            $ref: "#/definitions/Medico"
      operationId: createMedicos
      responses: 
        "201":
          description: success
          schema:
            $ref: "#/definitions/MedicoResponse"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse"
  /medicos/{id}:
    x-swagger-router-controller: epivector.medico.controller
    
    get: 
      description: Retorna una entidad
      parameters:
        - name: id
          description: Identificador de la entidad
          type: string
          in: path
          required: true
      operationId: getMedicosById
      responses:
        "200":
          description: success
          schema:
           $ref: "#/definitions/MedicoResponse"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse"
            
    put:
      description: Actualiza una entidad
      parameters:
        - name: id
          description: Identificador de la entidad
          type: string
          in: path
          required: true
        - in: body
          name: entidad
          schema:
            $ref: "#/definitions/Medico"
      operationId: updateMedicos
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/MedicoResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    
    delete:
      description: Borra una entidad
      parameters:
        - name: id
          description: Identificador de la entidad
          type: string
          in: path 
          required: true
      operationId: deleteMedicos
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/DeleteResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"                
...
  /fichagenerals:
    x-swagger-router-controller: epivector.fichageneral.controller
    
    get:
      operationId: getFichaGeneral
      description: Retorna la lista de entidades
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/fichageneralList"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse" 
            
    post:
      description: Agrega una nueva entidad
      parameters:
        - in: body
          name: entidad
          schema:
            $ref: "#/definitions/fichageneral"
      operationId: createFichaGeneral
      responses: 
        "201":
          description: success
          schema:
            $ref: "#/definitions/fichaResponse"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse"
  /fichageneral/{id}:
    x-swagger-router-controller: epivector.fichageneral.controller
    
    get: 
      description: Retorna una entidad
      parameters:
        - name: id
          description: Identificador de la entidad
          type: string
          in: path
          required: true
      operationId: getFichaGeneralId
      responses:
        "200":
          description: success
          schema:
           $ref: "#/definitions/fichaResponse"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse"            
            
            
    put:
      description: Actualiza una entidad
      parameters:
        - name: id
          description: Identificador de la entidad
          type: string
          in: path
          required: true
        - in: body
          name: entidad
          schema:
            $ref: "#/definitions/fichageneral"
      operationId: updateFichaGeneral
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/fichaResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"      
            
    delete:
      description: Borra una entidad
      parameters:
        - name: id
          description: Identificador de la entidad
          type: string
          in: path 
          required: true
      operationId: deleteFichaGeneral
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/DeleteResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"                
...
definitions:
...
  ErrorResponse:
    required:
      - message
    properties:
      code:
        type: number
      message:
        type: string
      description:
        type: string
  #  GeneralResponse:
  #    type: object
  #    properties:
  #      success:
  #        type: number
  #        description: returns 1 if successful
  #      description:
  #        type: string
  #        description: a short comment
  #    required:
  #      - success
  #      - description
...
  Medico:
    type: object
    properties:
      nombre: 
        type: string
        description: El nombre o nombres del medico
      apellido_p:
        type: string
        description: Apellido paterno del medico
      apellido_m:
        type: string
        description: Apellido materno del medico
      unidad_medica:
        type: number
        description: El id de la unidad medica
  MedicoResponse:
    type: object
    properties:
      id: 
        type: number
        description: Identifier
      nombre: 
        type: string
        description: El nombre o nombres del medico
      apellido_p:
        type: string
        description: Apellido paterno del medico
      apellido_m:
        type: string
        description: Apellido materno del medico
      unidad_medica:
        type: number
        description: El id de la unidad medica
  MedicoList:
    properties:
      medicos:
        type: array
        items:
          $ref: "#/definitions/MedicoResponse"
    required:
      - medicos
...
  DeleteResponse:
    type: object
    properties:
      success:
        type: number
        description: Retorna 1 en caso de exito
      description:
        type: string
        description: Comentario
    required:
      - success
      - description
  fichageneral:
    type: object
    properties:
      folio:
        type: number
        description: El folio  de la persona
      consultorio:
        type: number
        description: el comsultorio a asignar ala persona
      turno:
        type: number
        description: el turno a asignar
      nss:
        type: number
        description: numero de seguro social
      agregado:
        type: integer
        description: agregado
      apellido_p:
        type: string
        description: apellido de la persona
      apellido_m:
        type: string
        description: apellido materno
      nombres:
        type: string
        description: nombre de la persona
      fechadeNacimiento:
        type: string
        description: fecha nacimiento
      numeroExterior:
        type: number
        description: numero exterior
      numeroInterior: 
        type: number
        description: numero interior
      calle:
        type: string
        description: calle
      Entrecalle:
        type: string
        description: entre calle
      Ycalle:
        type: string
        description: y calle
      localidad:
        type: string
        description: localidad
      municipio:
        type: string
        description: municipio
      estado:
        type: string
        description: estado
      codigoPostal:
        type: number
        description: codigo
      edad:
        type: number
        description: edad
      peso:
        type: number
        description: peso
      estatura:
        type: number
        description: estatura
  fichaResponse:

    type: object
    properties:
      id:
        type: number
        description: identifier
      folio:
        type: number
        description: El folio  de la persona
      consultorio:
        type: number
        description: el comsultorio a asignar ala persona
      turno:
        type: number
        description: el turno a asignar
      nss:
        type: number
        description: numero de seguro social
      agregado:
        type: integer
        description: agregado
      apellido_p:
        type: string
        description: apellido de la persona
      apellido_m:
        type: string
        description: apellido materno
      nombres:
        type: string
        description: nombre de la persona
      fechadeNacimiento:
        type: string
        description: fecha nacimiento
      numeroExterior:
        type: number
        description: numero exterior
      numeroInterior: 
        type: number
        description: numero interior
      calle:
        type: string
        description: calle
      Entrecalle:
        type: string
        description: entre calle
      Ycalle:
        type: string
        description: y calle
      localidad:
        type: string
        description: localidad
      municipio:
        type: string
        description: municipio
      estado:
        type: string
        description: estado
      codigoPostal:
        type: number
        description: codigo
      edad:
        type: number
        description: edad
      peso:
        type: number
        description: peso
      estatura:
        type: number
        description: estatura
  fichageneralList:
    properties:
      fichageneral:
        type: array
        items:
          $ref: "#/definitions/fichaResponse"
    required:
      - fichageneral
...